backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pymongo==4.6.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.0
pydantic-settings==2.2.0
httpx==0.27.0
websockets==12.0